refactored each detail view section component into its own file
